SVN-Revision: 1366

+++ src/hpc/hpc-server/hpc-app-service-impl/src/main/java/gov/nih/nci/hpc/service/impl/HpcNotificationServiceImpl.java
@@ -12,11 +12,13 @@ package gov.nih.nci.hpc.service.impl;
 
 import static gov.nih.nci.hpc.service.impl.HpcDomainValidator.isValidNotificationSubscription;
 import gov.nih.nci.hpc.dao.HpcNotificationDAO;
+import gov.nih.nci.hpc.domain.datatransfer.HpcDataTransferDownloadStatus;
 import gov.nih.nci.hpc.domain.error.HpcErrorType;
 import gov.nih.nci.hpc.domain.notification.HpcEvent;
 import gov.nih.nci.hpc.domain.notification.HpcEventPayloadEntry;
 import gov.nih.nci.hpc.domain.notification.HpcEventType;
 import gov.nih.nci.hpc.domain.notification.HpcNotificationDeliveryMethod;
+import gov.nih.nci.hpc.domain.notification.HpcNotificationDeliveryReceipt;
 import gov.nih.nci.hpc.domain.notification.HpcNotificationSubscription;
 import gov.nih.nci.hpc.exception.HpcException;
 import gov.nih.nci.hpc.service.HpcNotificationService;
@@ -26,6 +28,8 @@ import java.util.EnumMap;
 import java.util.List;
 import java.util.Map;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 
 /**
@@ -40,6 +44,16 @@ import org.springframework.beans.factory.annotation.Autowired;
 public class HpcNotificationServiceImpl implements HpcNotificationService
 {
     //---------------------------------------------------------------------//
+    // Constants
+    //---------------------------------------------------------------------//
+	
+    // Event payload entries attributes.
+	private static final String DATA_TRANSFER_REQUEST_ID_ATTRIBUTE = 
+			                    "DATA_TRANSFER_REQUEST_ID";
+	private static final String DATA_TRANSFER_DOWNLOAD_STATUS_ATTRIBUTE = 
+                                "DATA_TRANSFER_DOWNLOAD_STATUS";
+	
+    //---------------------------------------------------------------------//
     // Instance members
     //---------------------------------------------------------------------//
 
@@ -50,6 +64,9 @@ public class HpcNotificationServiceImpl implements HpcNotificationService
     // The Notification DAO instance.
 	@Autowired
     private HpcNotificationDAO notificationDAO = null;
+	
+	// The logger instance.
+	private final Logger logger = LoggerFactory.getLogger(this.getClass().getName());
 	
     //---------------------------------------------------------------------//
     // Constructors
@@ -152,47 +169,93 @@ public class HpcNotificationServiceImpl implements HpcNotificationService
     }
     
     @Override
-    public void sendNotification(HpcEvent event, 
-                                 HpcNotificationDeliveryMethod deliveryMethod) 
-                                throws HpcException
+    public boolean sendNotification(String userId, HpcEventType eventType, 
+                                    List<HpcEventPayloadEntry> payloadEntries,
+                                    HpcNotificationDeliveryMethod deliveryMethod) 
     {
+    	// Input validation.
+    	if(userId == null || eventType == null || deliveryMethod == null) {
+    	   return false;
+    	}
+    	
     	// Locate the notification sender for this delivery method.
     	HpcNotificationSender notificationSender = notificationSenders.get(deliveryMethod);
     	if(notificationSender == null) {
-    	   throw new HpcException("Could not locate notification sender for: " + deliveryMethod,
-    			                  HpcErrorType.UNEXPECTED_ERROR);
+    	   logger.error("Could not locate notification sender for: " + deliveryMethod);
+    	   return false;
     	}
 
-    	notificationSender.sendNotification(event);
-    }
-    
-    @Override
-    public void createDeliveryReceipts(HpcEvent event) throws HpcException
-    {
-    	// TODO: Create delivery receipts, then delete the event.
-    	notificationDAO.deleteEvent(event.getId());
+    	// Send the notification.
+    	try {
+    	     notificationSender.sendNotification(userId, eventType, payloadEntries);
+    	     
+    	} catch(HpcException e) {
+    		    logger.error("failed to deliver event", e);
+    		    return false;
+    	}
     	
+    	return true;
     }
     
     @Override
-    public void addDataTransferDownloadCompletedEvent(String userId,
-    		                                          String dataTransferRequestId) throws HpcException
+    public void createNotificationDeliveryReceipt(String userId,
+    		                                      int eventId, 
+                                                  HpcNotificationDeliveryMethod deliveryMethod,
+                                                  boolean deliveryStatus)
+    {
+    	if(userId == null || deliveryMethod == null) {
+    	   return;
+    	}
+    	
+    	HpcNotificationDeliveryReceipt deliveryReceipt = new HpcNotificationDeliveryReceipt();
+    	deliveryReceipt.setUserId(userId);
+    	deliveryReceipt.setEventId(eventId);
+    	deliveryReceipt.setNotificationDeliveryMethod(deliveryMethod);
+    	deliveryReceipt.setDeliveryStatus(deliveryStatus);
+    	deliveryReceipt.setDelivered(Calendar.getInstance());
+    	
+    	try {
+    	     notificationDAO.upsertDeliveryReceipt(deliveryReceipt);
+    	     
+    	} catch(HpcException e) {
+    		    logger.error("Failed to create a delivery receipt", e);
+    	}
+    }
+    
+    @Override
+    public void archiveEvent(HpcEvent event)
+    {
+    	if(event == null) {
+    	   return;
+    	}
+    	
+    	// Delete the event from the active table and insert to the history.
+    	try {
+    	     notificationDAO.deleteEvent(event.getId());
+    	     notificationDAO.insertEventHistory(event);
+    	     
+    	} catch(HpcException e) {
+    		    logger.error("Failed to archive event", e);
+    	}
+    }
+    
+    @Override
+    public void addDataTransferDownloadCompletedEvent(
+    		       String userId, String dataTransferRequestId,
+    		       HpcDataTransferDownloadStatus dataTransferDownloadStatus) throws HpcException
     {
     	// Construct the event.
-    	HpcEventPayloadEntry payloadEntry = new HpcEventPayloadEntry();
-    	payloadEntry.setAttribute("DATA_TRANSFER_REQUEST_ID");
-    	payloadEntry.setValue(dataTransferRequestId);
-    	
     	HpcEvent event = new HpcEvent();
-    	event.setUserId(userId);
+    	event.getUserIds().add(userId);
     	event.setType(HpcEventType.DATA_TRANSFER_DOWNLOAD_COMPLETED);
-    	event.getPayloadEntries().add(payloadEntry);
-    	event.setCreated(Calendar.getInstance());
+    	event.getPayloadEntries().add(toPayloadEntry(DATA_TRANSFER_REQUEST_ID_ATTRIBUTE, 
+    			                                     dataTransferRequestId));
+    	event.getPayloadEntries().add(toPayloadEntry(DATA_TRANSFER_DOWNLOAD_STATUS_ATTRIBUTE, 
+    			                                     dataTransferDownloadStatus.value()));
 
     	// Persist to DB.
     	addEvent(event);
     }
-    
     
     //---------------------------------------------------------------------//
     // Helper Methods
@@ -208,14 +271,34 @@ public class HpcNotificationServiceImpl implements HpcNotificationService
     private void addEvent(HpcEvent event) throws HpcException
     {
     	// Input validation.
-    	if(event == null || event.getUserId() == null ||
-    	   event.getType() == null || event.getCreated() == null) {
+    	if(event == null || event.getUserIds() == null || event.getUserIds().isEmpty() ||
+    	   event.getType() == null) {
     	   throw new HpcException("Invalid event",
     			                  HpcErrorType.INVALID_REQUEST_INPUT);
     	}
 
+    	// Set the created timestamp.
+    	event.setCreated(Calendar.getInstance());
+    	
     	// Persist to DB.
     	notificationDAO.insertEvent(event);	
+    }
+    
+    /**
+     * Instantiate a payload entry object.
+     * 
+     * @param attribute The payload entry attribute.
+     * @param value The payload entry value
+     */
+    
+    private HpcEventPayloadEntry toPayloadEntry(String attribute, String value)
+    {
+		// Construct the event.
+		HpcEventPayloadEntry payloadEntry = new HpcEventPayloadEntry();
+		payloadEntry.setAttribute(attribute);
+		payloadEntry.setValue(attribute);
+		
+		return payloadEntry;
     }
 }
